some basic comparsion
