+++ crypto_examples/falcon_verify/compare.c
@@ -1,0 +1,27 @@
+#include <stddef.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <assert.h>
+
+int
+PQCLEAN_FALCON512_CLEAN_verify_raw(const uint16_t *c0, const int16_t *s2,
+                                   uint16_t *h, unsigned logn, uint8_t *tmp);
+
+#define FALCON_KEYGEN_TEMP_9    14336
+
+
+int main(void) {
+    uint8_t tmp[FALCON_KEYGEN_TEMP_9];
+
+    uint16_t buff_c0[] = {6673,10315,10940,3299,4987,5662,11800,9550,2694,7285,10708,9271,5302,10252,6550,4379,4864,610,7943,7824,5114,1447,6056,10284,11092,134,9036,3609,1300,12158,3845,9665,9733,1286,2694,3234,11398,5564,3406,1286,1931,8562,9165,1804,5996,11429,10552,6082,5037,7371,1689,12099,114,4137,5378,10635,6542,949,7802,9975,7195,4071,5425,9455,7531,139,11058,3459,6055,3991,11095,3331,9444,2786,2748,12130,10999,3562,7589,11383,11600,7646,8820,6571,13,1737,12199,5000,11779,4521,9776,3389,10782,8206,151,4370,7577,5380,501,6168,8233,9495,4724,3681,10087,8825,639,331,2216,5575,578,12232,6630,2559,3943,6350,1147,1979,7196,12003,1990,1946,11435,4858,1823,2570,5424,4417,10835,230,7174,6230,3965,7457,7365,6346,11573,7818,5093,10342,3646,11570,10541,1032,1714,11984,1599,2658,7471,3183,4841,1487,6818,6149,59,4224,11243,8476,10957,4761,6515,4821,4695,8327,3346,816,8069,10736,9283,7802,9985,12079,2597,9995,4482,6078,5882,3901,10058,10522,2821,2967,2082,9161,2133,2091,12077,3077,2426,7337,8708,6529,4377,5990,11895,5074,11521,8196,3455,7557,8214,6193,225,6541,5001,8846,4592,7685,4887,1958,4335,11441,5325,5959,802,7885,4183,2282,8597,4655,6144,2393,3354,1892,4025,5517,8980,5598,3140,3632,769,351,205,11781,2038,4797,271,9645,9006,681,8735,1765,5566,4225,10729,8154,12210,11994,8839,3632,7146,3860,1842,6456,2263,3770,2108,7291,2141,8070,8768,5603,12278,8039,3818,7992,6229,10915,2514,9612,9591,3061,1820,9580,12023,137,4519,5137,9208,4363,7112,2069,8425,11266,10784,12254,7314,14,6773,10897,6175,3446,8319,131,5493,9103,3454,10593,1849,10906,6904,3968,5873,913,1569,6431,7643,1451,3955,5784,5521,7948,5371,549,4727,6231,10299,980,11328,4806,619,182,1241,2903,1555,9013,9358,3462,4332,839,3895,469,5210,7597,7962,9497,10476,7109,2908,6482,5191,2153,6025,7581,7426,4571,4299,3812,3498,7244,4274,2022,8859,1670,12263,7258,7447,6788,2368,3473,10732,11379,11349,4244,144,11120,3583,11084,10320,7677,3648,11926,1727,5563,4196,9718,6620,959,1427,10017,4524,5386,1855,8744,3044,8480,9773,10804,7236,10273,2600,9864,2883,12040,5423,11087,8480,3826,8325,2103,2023,5979,9077,5463,7261,7750,10123,6595,7930,3869,12113,11,7509,4414,9764,5654,2063,592,4909,7381,11924,5296,678,11707,9237,5175,11937,2000,1783,6735,7838,6899,6358,8466,10072,4448,431,11900,7165,11098,1484,6558,9408,1946,12235,2949,9597,5259,11990,4477,10820,954,2897,3795,12120,7465,4025,10707,2944,8986,9069,8695,11418,7097,3889,995,2041,8724,12215,942,2300,11281,10910,8256,8645,5329,5550,4095,8297,6398,1056,5408,4645,1146,6612,12199,2656,6493,4550,5902,6539,2743,7884,11490,3665,6152,5200,556,6490,571,794,9852,1811,5973,10142,3747,9328,3682,7614,4133,7798,3711};
+    uint16_t buff_h[] = {308,5674,2938,937,513,10228,10316,6784,3124,10820,1141,4726,12037,11917,11696,2060,2858,10057,9324,2319,2888,5145,371,5961,9647,11284,4746,7851,8011,7430,7519,10929,2904,596,10283,10505,3605,11360,3900,4296,6396,5344,2473,11273,4427,10534,6248,5018,9514,11213,11410,10591,10688,4771,7718,11389,6483,7925,1610,8049,98,3983,7269,1263,1294,11419,402,3386,9792,8966,2191,1697,135,4170,11940,1255,7542,5118,8403,3113,8999,3774,3450,11821,7942,1541,11389,7729,10750,12163,4382,9027,8877,12236,11169,2378,7923,3740,7507,7693,6020,5349,8301,1283,11198,1685,4745,1141,368,4269,7860,12200,6433,1943,749,6428,6449,5457,7078,2430,6211,6731,1848,4943,10122,6667,2740,3668,9643,4512,6103,1199,8768,670,5246,6441,7479,4692,6742,11512,2859,11844,7075,7202,11946,5616,3872,11261,8679,11798,5728,8072,2773,5138,2052,6537,5566,4719,8474,8630,11562,7286,10423,5567,4631,1875,9842,11844,6385,8223,11709,3788,9957,783,4445,3084,6571,11340,10118,5374,7923,10385,9753,3522,5145,5827,535,12169,2679,6849,9451,8172,6463,7946,7227,2909,2713,9886,2775,9894,7635,9709,4547,4599,129,3773,5869,11848,8187,6147,8436,7398,5721,3849,9164,3697,4975,3479,7509,1618,5712,6599,3697,189,3953,9253,7503,7633,6416,8057,1428,221,9371,8131,1888,11280,8952,9997,11818,8758,8760,10835,267,11251,1360,11675,417,6464,3134,4128,8966,10097,5584,3967,10668,5433,7370,3552,4568,5922,12184,9928,9256,10187,7470,7895,6469,10796,8526,8869,8525,4655,8865,9896,9514,2426,12121,6146,10130,6862,6293,6967,9197,8375,9187,7781,5827,147,8572,11116,10371,8106,9270,3685,8636,533,8596,6148,12270,9751,3693,3460,9934,9899,2857,9816,2526,8895,11004,2559,7380,7372,1741,4345,10756,3202,148,10513,11460,649,3081,11077,774,3907,4145,6951,11026,6090,11338,11340,6319,3316,5232,10784,6202,629,4776,959,11773,6858,1944,6529,7290,10946,2938,7897,5582,11495,8738,9329,5117,4659,8927,11602,2969,7196,1799,4967,6529,6382,9194,5136,4863,2179,4237,6763,2864,7202,11662,3419,5327,10109,4404,662,11953,1663,9576,11623,11832,6337,5235,4237,12262,5452,8358,1908,3727,3960,2164,2351,3679,2455,5285,10959,11242,7442,3629,5474,4908,2905,6379,3080,1658,10599,10743,11096,12063,10721,9675,4995,11614,1645,6427,3228,5132,10619,6031,1508,5885,5535,66,1672,11467,4029,5678,9628,7434,9673,8710,8657,3637,6590,1454,11765,846,2792,7187,5235,3005,6414,11054,7313,5880,528,369,9388,11768,1215,2590,7761,484,2214,11229,1028,5115,8868,12266,10812,7880,5108,8717,9361,3242,7185,1069,2801,10923,8266,10663,8701,187,3233,2343,8827,8903,10803,2006,10098,1306,11329,5315,11430,7683,6331,11304,1334,4542,4513,9917,1260,9488,2105,950,2622,287,11985,1009,5567,8205,2044,5968,8808,6339,4580,10256,2134,1182,3976,1977,50,2688,1143};
+    int16_t buff_s2[] = {-211,39,56,180,-287,-64,-142,-178,-197,-53,262,196,86,0,-100,124,-116,-38,152,155,34,226,114,105,69,329,26,28,-92,35,-131,17,124,-298,-159,-140,168,-157,-181,2,174,58,-142,32,-142,-280,-95,-70,215,-277,-157,-43,17,2,448,-30,557,-93,12,285,-163,61,-114,125,188,-139,160,203,-423,67,-187,0,273,82,-6,-212,186,167,41,96,238,139,9,-97,-246,-89,29,68,172,-100,175,-149,-137,-66,121,-30,2,-62,-140,277,-18,280,70,-153,385,-356,484,29,74,145,138,53,-141,184,136,57,126,-140,132,-12,-224,222,-199,-140,44,-182,58,-194,24,-263,-5,-148,47,138,324,-143,182,-124,-32,-57,79,161,84,-130,116,228,-14,-114,-25,89,-120,129,264,11,67,-110,-162,-229,-244,-27,7,-209,139,-157,65,-102,-55,123,-295,-258,-104,-145,77,-7,-79,-30,-77,226,-121,2,-139,-29,-161,198,-139,3,-194,-49,179,113,-81,-2,-41,-118,-61,-208,-67,28,150,-169,-145,304,33,22,159,255,-84,-182,-376,264,57,206,65,44,169,-26,51,-306,-200,-183,-436,214,226,-56,-71,43,-207,136,67,319,-48,-130,132,117,110,-326,-13,-169,-48,-176,-314,-142,-217,-100,-13,-38,-26,-294,-47,227,89,175,-173,42,117,-93,-47,-29,19,100,-61,-185,-202,187,66,40,72,50,-12,-16,-161,238,8,-150,-17,-168,-36,80,47,-89,98,-179,117,317,-261,376,18,-40,10,44,-115,121,174,-361,214,-111,129,-63,-214,471,23,-128,199,-95,-184,-198,74,-127,160,-336,76,-159,-19,-232,-9,16,128,175,49,-91,68,-180,-42,-51,-53,-43,84,155,53,190,-96,-113,-66,-42,-230,134,249,-95,-280,-25,-123,5,-176,310,-195,149,53,197,38,-150,300,-13,62,-73,160,-185,52,-168,-33,74,310,-83,28,-338,-148,-41,115,-64,124,-6,-135,-37,-23,-92,132,-82,-122,123,236,35,57,198,159,133,-127,139,166,-95,-142,-102,-39,32,-171,119,14,177,10,128,219,5,1,27,-40,191,-66,-118,-116,41,-218,-46,276,-27,-145,86,-162,86,-156,120,58,103,173,-65,-175,45,-7,-146,158,53,15,-212,-111,223,503,-21,214,85,179,82,61,-84,77,98,106,278,-23,169,-119,-443,-3,220,98,141,250,7,37,-36,99,58,-88,46,130,-305,161,269,-25,47,-260,94,28,146,-252,-123,-292,-65,-68,-244,-238,87,-229,-125,216,-38,-81,288,-225,-135,300,-191,-84,214,-133,62,-78,438,114,-74,-242,-182,142,-72,-244,4,-186,-68,-192,-191,335,52,85,97,91,16};
+    
+    int dummy;
+
+    for (int i = 0; i < 2000000; i++) {
+        dummy |= PQCLEAN_FALCON512_CLEAN_verify_raw(buff_c0, buff_s2, buff_h, 9, tmp);
+    }
+
+    return dummy;
+}
